requirements.txt
Flask
flask-mysqldb
passlib
PyYAML